med-control-back/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Base de datos
supabase>=2.10.0

# HTTP client
httpx>=0.25.0

# LLM providers
ollama==0.1.6

# Environment
python-dotenv==1.0.1

# Logging
structlog==24.1.0

# Validación
pydantic==2.5.3
pydantic-settings==2.1.0

# OCR — procesamiento de recetas médicas
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0

# Web search — búsqueda de información médica
tavily-python==0.3.3

# Auth
PyJWT>=2.8.0
bcrypt>=4.0.0

# Utilidades
pyfiglet==1.0.2
